APPS-1121 Update integration test suite (feat) RegisteredTest.py - tested bas class
