+++ dxcint/tests/test_RegisteredTest.py
@@ -16,12 +16,13 @@ def registered_test_wdl(fixtures_dir, context_init):
     )
 
 
-@pytest.fixture(scope="session")
-def build_executable_wdl(registered_test_wdl):
-    _ = registered_test_wdl.exec_id()
+@pytest.fixture(scope="function")
+def build_executable_wdl(registered_test_wdl, terraform_init, change_to_root_dir):
+    _ = terraform_init.build()
+    _ = registered_test_wdl.exec_id
     yield registered_test_wdl
     try:
-        container_remove_objects(registered_test_wdl.project_id, {"objects": [registered_test_wdl.exec_id]})
+        container_remove_objects(registered_test_wdl.context.project_id, {"objects": [registered_test_wdl.exec_id]})
     except DXAPIError as e:
         print(f"Could not find {registered_test_wdl.exec_id}. Exiting")
         print(e)
