+++ dxcint/tests/test_Terraform.py
@@ -6,13 +6,6 @@ from dxcint.Terraform import Terraform
 from dxcint.Dependency import DependencyFactory
 
 
-@pytest.fixture(scope="session")
-def terraform_init(dependency_conf_dir, context_init):
-    dependency_factory = DependencyFactory(os.path.join(dependency_conf_dir, "mock_dependency_src.json"))
-    bin_dependency = dependency_factory.make()
-    yield Terraform(languages={"wdl"}, context=context_init, dependencies={bin_dependency})
-
-
 def test__build_compiler(change_to_root_dir, terraform_init):
     compiler_built = terraform_init._build_compiler()
     assert compiler_built
